# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documents",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/receiver/create": {
            "post": {
                "tags": [
                    "Receiver"
                ],
                "summary": "Create new receiver",
                "operationId": "822ebe838d9f319ee147ef866f280372",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account_no",
                                    "firstname",
                                    "lastname",
                                    "phone_number"
                                ],
                                "properties": {
                                    "account_no": {
                                        "type": "string",
                                        "example": "0234245345"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "Receiver"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "A"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0812221111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Create new receiver successfully"
                                        },
                                        "receiver": {
                                            "properties": {
                                                "account_no": {
                                                    "type": "string",
                                                    "example": "0234245345"
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "Receiver A"
                                                },
                                                "create_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-15 14:23:45"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Create failed"
                    }
                }
            }
        },
        "/api/receiver/list": {
            "post": {
                "tags": [
                    "Receiver"
                ],
                "summary": "Receiver List",
                "operationId": "0f27ab282282b2ef8e905cac2fad2741",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "keyword": {
                                        "description": "Search by account_no, phone_number, firstname, lastname, fullname",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "sort": {
                                        "description": "Sort by firstname => value: asc,desc",
                                        "type": "string",
                                        "example": "asc"
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "limit": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 12
                                                }
                                            },
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "account_no": {
                                                        "type": "string",
                                                        "example": "0234245345"
                                                    },
                                                    "firstname": {
                                                        "type": "string",
                                                        "example": "Receiver"
                                                    },
                                                    "lastname": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "0812221111"
                                                    },
                                                    "create_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-05-15 14:23:45"
                                                    },
                                                    "update_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-05-15 14:23:45"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Not found data"
                    }
                }
            }
        },
...
    },
...
}